requirements.txt
# Core dependencies
Flask==3.1.0
onvif-zeep==0.2.12
opencv-python==4.11.0.86

# Flask dependencies
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.1.8
itsdangerous==2.2.0
blinker==1.9.0

# ONVIF dependencies
zeep==4.3.1
requests==2.32.3
urllib3==2.3.0
lxml==5.3.2

# Other utilities
python-dateutil==2.8.2
six==1.16.0 